+++ src/Wsdl/WebServiceFactory.php
@@ -3,25 +3,36 @@ declare(strict_types = 1);
 
 namespace Skaut\Skautis\Wsdl;
 
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Skaut\Skautis\InvalidArgumentException;
 
-/**
- * @inheritdoc
- */
-class WebServiceFactory implements WebServiceFactoryInterface
+final class WebServiceFactory implements WebServiceFactoryInterface
 {
 
-    /** @var string Třída webové služby */
-    protected $class;
+    /**
+     * @var string Třída webové služby
+     */
+    private $class;
+
+    /**
+     * @var EventDispatcherInterface|null
+     */
+    private $eventDispatcher;
 
 
-    public function __construct(string $className = WebService::class)
-    {
-       if (!is_a($className, WebServiceInterface::class, true)) {
-         throw new InvalidArgumentException("Argument must be class name of a class implementing WebServiceInterface. '$className' given");
-       }
+    /**
+     * @param string $className Constructor must accept SoapClient, SOAP options array and EventDispatcherInterface
+     */
+    public function __construct(
+      string $className = WebService::class,
+      ?EventDispatcherInterface $eventDispatcher = null
+    ) {
+        if (!is_a($className, WebServiceInterface::class, true)) {
+          throw new InvalidArgumentException("Argument must be class name of a class implementing WebServiceInterface. '$className' given");
+        }
 
         $this->class = $className;
+        $this->eventDispatcher = $eventDispatcher;
     }
 
     /**
@@ -34,6 +45,6 @@ class WebServiceFactory implements WebServiceFactoryInterface
         }
 
         $soapClient = new \SoapClient($url, $options);
-        return new $this->class($soapClient, $options);
+        return new $this->class($soapClient, $options, $this->eventDispatcher);
     }
 }